requirements.txt
pandas==1.3.5
airflow-dbt==0.4.0
dbt-core==1.7.7
dbt-snowflake==1.7.1
snowflake-connector-python==3.7.0
snowflake-connector-python[pandas]